javaCore / Project tree naming fix
