+++ documentation/javaCore/lesson_7_collections/collections1/Task1&2_Chats.txt
@@ -1,0 +1,20 @@
+Задание 1
+Дан список чатов.
+Каждый чат состоит из двух полей: название и количество пользователей в этом чате.
+
+Задача:
+
+- Удалить с помощью итератора из этого списка те чаты, что содержат менее 1000 пользователей.
+
+- Оставшиеся чаты отсортировать с помощью компараторов по убыванию по количеству пользователей, а если это количество совпадает, то по названию в алфавитном порядке.
+
+- Также предоставить сортировку чатов по названию по умолчанию.
+
+Задание 2
+Дан список чатов с предыдущего задания, только вместо поля “количество пользователей” будет список объектов типа Пользователь, который имеет 3 поля: id (целочисленный идентификатор), имя и возраст.
+
+Задача:
+
+- Преобразовать список чатов в один список пользователей всех чатов, возраст которых больше 18 лет
+
+- С помощью итератора посчитать средний возраст всех оставшихся пользователей.
