+++ documentation/javaCore/lesson_7_collections/collections2/Task1_Cinema.txt
@@ -1,0 +1,27 @@
+Задание 1
+
+Даны 2 класса:
+
+- Фильм с 5-ю полями: Уникальный Целочисленный идентификатор (id), Год издания, Месяц издания, Жанр и Рейтинг;
+
+- Кинотеатр, где есть всего лишь одно единственное поле: отсортированный ассоциативный массив, где ключом является год издания, а значением - все фильмы, выпустившиеся в этом году.
+
+Добавить функционал в кинотеатр таким образом, чтобы можно было:
+
+- добавлять в него новый фильм
+
+- получить все фильмы по переданному году
+
+- получить все фильмы по переданному году и месяцу
+
+- получить все фильмы по переданному жанру
+
+- получать ТОП-10 фильмов отсортированные по Рейтингу в порядке убывания
+
+Учесть следующее:
+
+- в кинотеатре фильмы должны храниться в порядке их добавления в него (т.е. предусмотреть порядок значения ассоциативного массива)
+
+- не должен добавляться фильм, если такой уже есть в кинотеатре
+
+Продемонстрировать работу кинотеатра в отдельном классе.
