+++ src/main/java/org/dmdev1/javaCore/lesson_4_strings/homework_Task1/StringFilter1_v1.java
@@ -1,0 +1,28 @@
+package org.dmdev1.javaCore.lesson_4_strings.homework_Task1;
+
+public class StringFilter1_v1 {
+
+    public static void main(String[] args) {
+        String str = "abc Cpddd Dio OsfWw";
+        String output = removeConsecutiveDuplicatesAndSpaces(str);
+        System.out.println(output);
+    }
+
+
+    private static String removeConsecutiveDuplicatesAndSpaces(String str) {
+        StringBuilder sb = new StringBuilder();
+        char prevChar = '\0';
+
+        for (char c : str.toCharArray()) {
+            if (c != ' ') {
+                char upperC = Character.toUpperCase(c);
+                if (upperC != prevChar) {
+                    sb.append(upperC);
+                }
+                prevChar = upperC;
+            }
+        }
+        return sb.toString().toUpperCase();
+
+    }
+}
