+++ src/main/java/org/dmdev1/javaCore/lesson_5_oop_concepts/Task2_Space/Asteroid.java
@@ -1,0 +1,14 @@
+package org.dmdev1.javaCore.lesson_5_oop_concepts.Task2_Space;
+
+public class Asteroid extends AstronomicalObject {
+
+
+    public Asteroid(String name, long diameterInKm, double massInSolarM) {
+        super(name, diameterInKm, massInSolarM);
+    }
+
+    @Override
+    public void print() {
+
+    }
+}
