--- src/main/java/org/dmdev1/oop_level2/oop_concepts/Task1_Atm/Input.java
@@ -1,13 +1,0 @@
-package org.dmdev1.oop_level2.oop_concepts.Task1_Atm;
-
-public class Input {
-
-    public static void main(String[] args) {
-        Atm_v1 atm = new Atm_v1(0, 0, 0);
-        System.out.println(atm.getAtmBalance());
-        atm.addBanknotes(0, 1, 1);
-        System.out.println(atm.getAtmBalance());
-        atm.cashWithdrawal(70);
-    }
-
-}
