--- src/main/java/org/dmdev1/oop_level2/oop_concepts/Task2_Space/Asteroid.java
@@ -1,14 +1,0 @@
-package org.dmdev1.oop_level2.oop_concepts.Task2_Space;
-
-public class Asteroid extends AstronomicalObject {
-
-
-    public Asteroid(String name, long diameterInKm, double massInSolarM) {
-        super(name, diameterInKm, massInSolarM);
-    }
-
-    @Override
-    public void print() {
-
-    }
-}
